Consolidate DefinitionList and DefinitionListItem
