+++ src/components/ComponentFactory.js
@@ -15,8 +15,8 @@ import Container from './Container';
 import CSSClass from './CSSClass';
 import CTA from './CTA';
 import CTABanner from './Banner/CTABanner';
-import DefinitionList from './DefinitionList';
-import DefinitionListItem from './DefinitionListItem';
+import DefinitionList from './DefinitionList/DefinitionList';
+import DefinitionListItem from './DefinitionList/DefinitionListItem';
 import DeprecatedVersionSelector from './DeprecatedVersionSelector';
 import Describe from './Describe';
 import DriversIndexTiles from './DriversIndexTiles';
